+++ src/ray.odin
@@ -8,7 +8,7 @@ Ray :: struct {
     direction: Point3
 }
 
-ray_at :: proc(ray: ^Ray, t: f32) -> Point3 { 
+ray_at :: proc(ray: ^Ray, t: f64) -> Point3 { 
     return ray.origin + t * ray.direction
 }
